OCD code inspections - boxing and unboxing

+++ src/main/java/org/pushingpixels/lafplugin/XMLElement.java
@@ -1273,7 +1273,7 @@ public class XMLElement
                 throw this.invalidValue(name, (String) key);
             }
         }
-        return result.intValue();
+        return result;
     }
 
 
@@ -1337,7 +1337,7 @@ public class XMLElement
             return defaultValue;
         } else {
             try {
-                return Double.valueOf(value).doubleValue();
+                return Double.valueOf(value);
             } catch (NumberFormatException e) {
                 throw this.invalidValue(name, value);
             }
@@ -1411,7 +1411,7 @@ public class XMLElement
                 throw this.invalidValue(name, (String) key);
             }
         }
-        return result.doubleValue();
+        return result;
     }
